MALGUARD/requirements.txt
flask
scikit-learn
pandas
joblib
tldextract
requests
python-dotenv
